GodsAssettoCorsaServerManager - Most Stable version
This version is the epitome of stability and resilience that I have been striving to achieve, a testament to the countless hours of work put in to ensure the best possible experience for all users.

As the main and most stable release to date, it forms the solid base for all future updates. It's from this version that I will consistently innovate and introduce new, more efficient versions. Each iteration will be a step forward, but this version will always serve as a reliable fallback, embodying the core functionalities and efficiency that GodsAssettoCorsaServerManager is all about.

I believe that this new release will bring immense value to your projects, and I look forward to receiving your feedback and continuously improving the tool based on your experience.